Remove typetag from UInt/SInt utils
Readd to internal/package.scala
